From: lars@segv.dk (Lars Munch) Hi
This patch adds two new functions, ecore_pipe_close_read and
ecore_pipe_close_write, to ecore_pipe. The purpose it to enable
ecore_pipe to be used together with fork (see example below).

The patch also handles if the read or write end of the pipe closes.



SVN revision: 40305