This closes bug #507
If e17 is compiled with tslib but during runtime the environment variable
TSLIB_TSDEVICE is not set, ecore will accidentally use the file descriptor 0
(probably stdin) for tslib.

The problem is that _ecore_fb_ts_fd is initialized to 0, which is BAD BAD BAD
for file descriptors. It should be initialized to -1. The attached patch fixes
this. 

Thanks to John Ogness <john.ogness@linutronix.de> for bug report and fix




SVN revision: 45703