From: Mike McCormack <mj.mccormack@samsung.com>
ecore_config code may end up closing stdin, due to assuming 0 is a valid fd.

I know this code is deprecated, and it's being kept around "for reference".

Either:
A) this bug should still be fixed so bad code is not copied elsewhere

OR

B) ecore_config should be deleted so bad code is not copied elsewhere

thanks,

Mike


SVN revision: 56275