From: Jihoon Kim <jihoon48.kim@samsung.com> Subject: [E-devel] [PATCH] Add Ecore_IMF API to set the attirbute of input panel
For supporting virtual keyboard, I'd like to add
ecore_imf_context_input_panel_enabled_set/get APIs. The detail description of
each API is included in the patch file as doxygen format.

If input panel is in 'enabled' status, the immodule will request to
show the input panel automatically When the input widget such as entry is
clicked or has focus. In some case, application programmers want to control
the input panel manually (not automatically), so I implement this API.



SVN revision: 60504