Ecore_Evas: Patch from Robert Bradford <robert.bradford@intel.com>:
The challenge here is that the native window representation is stored
in Ecore_Evas's prop.window. But currently there is no checking of
what driver the Ecore_Evas is for when calls are made to e.g.

ecore_evas_software_x11_window_get.

The attached change to Ecore makes the appropriate functions return 0
or NULL if the driver for the Ecore doesn't match as expected. This
can then be used to identify if an Ecore_Evas is e.g. from X11 or from
Wayland.



SVN revision: 71453