From: Mariusz Grzegorczyk <mariusz.g@samsung.com> Subject: [E-devel] Ecore: Save proper evas engine buffer depth when _ecore_evas_resize is done
I have found bug in ecore evas buffer. It is created by default as RGB32,
and then it can be changed by ecore_evas_alpha_set function, but when resize
is done it is always created as ARGB32.

Patch in attachment checks in resize if alpha channel is set, and then set
appropriate buffer depth.



SVN revision: 71537