Backport: e38c985cd0250814d2b222ce8ab9a599e81aeee4
ecore/wayland: Actually wait for the sync callback on ecore_wl_sync().

The wl_display_sync() request doesn't really wait for all the requests
to be processed, but instead sends a request for the "done" event.
Wayland relies on the fact that the requests are processed in order, so
when the "done" event is received, it means that all the other requests
requested prior to the respective "sync" have been processed already.

This commit makes the ecore_wl_sync() call actually wait for its "done"
event (thus blocking the ecore mainloop).