ecore_evas/wayland: Artificially call the changed_state callback.
On maximize and fullscreen set, call this callback manually, which will
make Elementary emit the maximized and fullscreen events, and also
update its internal state for these attributes.

The implementation is not done correctly as it is in the dev branch
because we are trying to not create new API's (even for the Wayland
backend, which is still under heavy development) on the stable branch,
unless it's extremely needed.