ecore_evas/wayland: Add a clipper to the frame object.
The frame object is a simple smart object with two children: a rectangle
(background) and a text (title). During the render phase, these objects
get wrongly clipped to the client area clipper, because they have no
clipper.

This commit adds a clipper to the frame smart object, and clip the other
children to it, so they won't get clipped to the client area clipper.

It fix an old bug where there was a flickering behavior on the window
decorations. The bug only happened on wayland_egl because of differences
on how this backend handles dirty areas, but it could be a source of new
bugs on wayland_shm too, so the solution is applied to both engines.