Ecore_con send error event when dns error.
This patch is a backport of http://git.enlightenment.org/core/efl.git/commit/?id=838c3bed85535559713e57f6e523da1f73c0ca31 :
Fixes a bug where when you try to connect to a server using an
inexistant DNS, or having a DNS resolution error, there isnt any event
generated for the application to know.

This is not a complete fix has you need to fix a timeout value for it to
happen, but still, it makes it possible to work around it.

More work is needed on this.

/hug tasn.