Put PeekNamedPipe back and add a sleep().
ReadFile is not able to detect when a pipe is closed.
Therefore, it will endlessly wait for data while the other side is dead.

This patch puts PeekNamedPipe back, but adds a 100ms sleep when it
returns a size of 0.

Doing so, were back to checking exe->close_threads that is the only way
to know if the other side is dead or not (PeekNamedPipe wont return any
error is the other side closed the pipe).

So as long as theres no data to read, were doing 100ms sleeps.

It isnt possible to use OVERLAPPED as we have anonymous pipes.
We should use named pipes to use OVERLAPPED, and we cant add them without
doing an ABI break.